[style] : 개설하기 로딩팝업 추가, 채팅방 배경화면 이미지 늘리기, 책 검색 시 로직 조금 변경
